refactor: 사용하지 않는 UI 컴포넌트 제거
