--- techpick/src/ui/Popover/Popover.tsx
@@ -1,30 +1,0 @@
-'use client';
-
-import { popoverStyle } from '@/ui/Popover/Popover.css';
-import * as PopoverPrimitive from '@radix-ui/react-popover';
-import * as React from 'react';
-
-const Popover = PopoverPrimitive.Root;
-
-const PopoverTrigger = PopoverPrimitive.Trigger;
-
-const PopoverContent = React.forwardRef<
-  React.ElementRef<typeof PopoverPrimitive.Content>,
-  React.ComponentPropsWithoutRef<typeof PopoverPrimitive.Content>
->(({ className, align = 'center', sideOffset = 4, ...props }, ref) => {
-  return (
-    // <PopoverPrimitive.Portal>
-    <PopoverPrimitive.Content
-      ref={ref}
-      align={align}
-      sideOffset={sideOffset}
-      className={`${className} ${popoverStyle}`}
-      {...props}
-    />
-    // </PopoverPrimitive.Portal>
-  );
-});
-
-PopoverContent.displayName = PopoverPrimitive.Content.displayName;
-
-export { Popover, PopoverTrigger, PopoverContent };
